Prework for Gregory Anderson
Included is my prework for the promises, node, and scraping workshop. I did not fully comprehend the scraping workshop and have spoken to some people and will be attempting to execute that tonight with new resources. Also, originally I forked this repo to my own github and then submitted a PR to myself. Admittedly, this sounds pretty junior varsity, but I swear it made sense when I did it in my head.